Moving and consolidating packages
