+++ components/data-sources/org.wso2.carbon.datasource.reader.cassandra/src/main/java/org/wso2/carbon/datasource/reader/cassandra/config/SocketOptionsConfig.java
@@ -15,7 +15,7 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.wso2.carbon.datasource.reader.cassandra.config.socket;
+package org.wso2.carbon.datasource.reader.cassandra.config;
 
 import com.datastax.driver.core.SocketOptions;
